requirements.txt
mcp>=1.0.0
netmiko>=4.0.0
ncclient>=0.6.0
scrapli[asyncssh]>=2024.0.0
requests>=2.28.0
pyats>=25.11
genie>=25.11
ntc-templates>=8.0.0
python-dotenv>=1.0.0

# Nornir (parallel automation)
nornir>=3.4.0
nornir-scrapli>=2024.0.0
nornir-netmiko>=1.0.0

# Ansible
ansible>=9.0.0
ansible-lint>=24.0.0

# ChatOps
discord.py>=2.3.0
aiohttp>=3.13.3  # CVE-2025-69223 through CVE-2025-69230
pynetbox

# Notebooks / Visualization
jupyter>=1.0.0
matplotlib>=3.8.0
pandas>=2.1.0
plotly>=5.18.0
networkx>=3.2.0

# Memory System
aiosqlite>=0.19.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# RAG Chatbot
anthropic>=0.18.0
pymupdf>=1.23.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Configuration
pydantic-settings>=2.0.0

# Dashboard API
Flask>=3.0.0
flask-cors>=4.0.0
flasgger>=0.9.7
flask-socketio>=5.3.0,<5.5.0
flask-limiter>=3.5.0
flask-caching>=2.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0

# Database
psycopg2-binary>=2.9  # PostgreSQL support (optional)
alembic>=1.13.0  # Schema migrations
sqlalchemy>=2.0.0  # Alembic requirement (app code stays raw SQL)

# Caching
redis>=5.0.0

# Monitoring
influxdb>=5.3.0  # Time-series database client

# Task Queue
celery>=5.3.0

# Production server
gunicorn>=21.0.0
gevent>=23.0.0
gevent-websocket>=0.10.1
defusedxml>=0.7.1

# Enterprise Security
hvac>=2.0.0  # HashiCorp Vault client
pyotp>=2.9.0  # TOTP MFA
qrcode[pil]>=7.4.0  # QR code generation for MFA
cryptography>=41.0.0  # Encryption for MFA secrets

# SNMP Support
pysnmp>=6.0.0
tenacity>=8.2.0  # Retry logic

# Testing
pytest-cov>=4.1.0

# Security pins (CVE fixes)
urllib3>=2.6.3  # CVE-2025-66418, CVE-2025-66471, CVE-2026-21441
werkzeug>=3.1.5  # CVE-2026-21860
python-multipart>=0.0.22  # CVE-2026-24486
pynacl>=1.6.2  # CVE-2025-69277
protobuf>=6.33.5  # CVE-2026-0994 (transitive dep)
filelock>=3.20.3  # CVE-2026-22701 (transitive dep)
apscheduler>=3.10.0
reportlab>=4.0.0
